+++ internal/impl/aws/output_kinesis.go
@@ -2,7 +2,6 @@ package aws
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"time"
 
@@ -12,105 +11,117 @@ import (
 	"github.com/aws/aws-sdk-go/service/kinesis/kinesisiface"
 	"github.com/cenkalti/backoff/v4"
 
-	"github.com/benthosdev/benthos/v4/internal/batch/policy"
-	"github.com/benthosdev/benthos/v4/internal/bloblang/field"
-	"github.com/benthosdev/benthos/v4/internal/bundle"
 	"github.com/benthosdev/benthos/v4/internal/component"
 	"github.com/benthosdev/benthos/v4/internal/component/output"
-	"github.com/benthosdev/benthos/v4/internal/component/output/batcher"
-	"github.com/benthosdev/benthos/v4/internal/component/output/processors"
-	"github.com/benthosdev/benthos/v4/internal/docs"
-	sess "github.com/benthosdev/benthos/v4/internal/impl/aws/session"
-	"github.com/benthosdev/benthos/v4/internal/log"
-	"github.com/benthosdev/benthos/v4/internal/message"
-	"github.com/benthosdev/benthos/v4/internal/old/util/retries"
+	"github.com/benthosdev/benthos/v4/internal/impl/aws/config"
+	"github.com/benthosdev/benthos/v4/internal/impl/pure"
+	"github.com/benthosdev/benthos/v4/public/service"
 )
+
+const (
+	// Kinesis Output Fields
+	koFieldStream       = "stream"
+	koFieldHashKey      = "hash_key"
+	koFieldPartitionKey = "partition_key"
+	koFieldBatching     = "batching"
+)
+
+type koConfig struct {
+	Stream       string
+	HashKey      *service.InterpolatedString
+	PartitionKey *service.InterpolatedString
+
+	session     *session.Session
+	backoffCtor func() backoff.BackOff
+}
+
+func koConfigFromParsed(pConf *service.ParsedConfig) (conf koConfig, err error) {
+	if conf.Stream, err = pConf.FieldString(koFieldStream); err != nil {
+		return
+	}
+	if conf.PartitionKey, err = pConf.FieldInterpolatedString(koFieldPartitionKey); err != nil {
+		return
+	}
+	if pConf.Contains(koFieldHashKey) {
+		if conf.HashKey, err = pConf.FieldInterpolatedString(koFieldHashKey); err != nil {
+			return
+		}
+	}
+	if conf.session, err = GetSession(pConf); err != nil {
+		return
+	}
+	if conf.backoffCtor, err = pure.CommonRetryBackOffCtorFromParsed(pConf); err != nil {
+		return
+	}
+	return
+}
+
+func koOutputSpec() *service.ConfigSpec {
+	return service.NewConfigSpec().
+		Stable().
+		Version("3.36.0").
+		Categories("Services", "AWS").
+		Summary(`Sends messages to a Kinesis stream.`).
+		Description(output.Description(true, true, `
+Both the `+"`partition_key`"+`(required) and `+"`hash_key`"+` (optional) fields can be dynamically set using function interpolations described [here](/docs/configuration/interpolation#bloblang-queries). When sending batched messages the interpolations are performed per message part.
+
+### Credentials
+
+By default Benthos will use a shared credentials file when connecting to AWS services. It's also possible to set them explicitly at the component level, allowing you to transfer data across accounts. You can find out more [in this document](/docs/guides/cloud/aws).`)).
+		Fields(
+			service.NewStringField(koFieldStream).
+				Description("The stream to publish messages to."),
+			service.NewInterpolatedStringField(koFieldPartitionKey).
+				Description("A required key for partitioning messages."),
+			service.NewInterpolatedStringField(koFieldHashKey).
+				Description("A optional hash key for partitioning messages.").
+				Optional().
+				Advanced(),
+			service.NewOutputMaxInFlightField().
+				Description("The maximum number of parallel message batches to have in flight at any given time."),
+			service.NewBatchPolicyField(koFieldBatching),
+		).
+		Fields(config.SessionFields()...).
+		Fields(pure.CommonRetryBackOffFields(0, "1s", "5s", "30s")...)
+}
+
+func init() {
+	err := service.RegisterBatchOutput("aws_kinesis", koOutputSpec(),
+		func(conf *service.ParsedConfig, mgr *service.Resources) (out service.BatchOutput, batchPolicy service.BatchPolicy, maxInFlight int, err error) {
+			if maxInFlight, err = conf.FieldMaxInFlight(); err != nil {
+				return
+			}
+			if batchPolicy, err = conf.FieldBatchPolicy(koFieldBatching); err != nil {
+				return
+			}
+			var wConf koConfig
+			if wConf, err = koConfigFromParsed(conf); err != nil {
+				return
+			}
+			out, err = newKinesisWriter(wConf, mgr)
+			return
+		})
+	if err != nil {
+		panic(err)
+	}
+}
 
 const (
 	kinesisMaxRecordsCount = 500
 	mebibyte               = 1048576
 )
 
-func init() {
-	err := bundle.AllOutputs.Add(processors.WrapConstructor(func(c output.Config, nm bundle.NewManagement) (output.Streamed, error) {
-		kin, err := newKinesisWriter(c.AWSKinesis, nm)
-		if err != nil {
-			return nil, err
-		}
-		w, err := output.NewAsyncWriter("aws_kinesis", c.AWSKinesis.MaxInFlight, kin, nm)
-		if err != nil {
-			return w, err
-		}
-		return batcher.NewFromConfig(c.AWSKinesis.Batching, w, nm)
-	}), docs.ComponentSpec{
-		Name:    "aws_kinesis",
-		Version: "3.36.0",
-		Summary: `
-Sends messages to a Kinesis stream.`,
-		Description: output.Description(true, true, `
-Both the `+"`partition_key`"+`(required) and `+"`hash_key`"+` (optional)
-fields can be dynamically set using function interpolations described
-[here](/docs/configuration/interpolation#bloblang-queries). When sending batched messages the
-interpolations are performed per message part.
-
-### Credentials
-
-By default Benthos will use a shared credentials file when connecting to AWS
-services. It's also possible to set them explicitly at the component level,
-allowing you to transfer data across accounts. You can find out more
-[in this document](/docs/guides/cloud/aws).`),
-		Config: docs.FieldComponent().WithChildren(
-			docs.FieldString("stream", "The stream to publish messages to."),
-			docs.FieldString("partition_key", "A required key for partitioning messages.").IsInterpolated(),
-			docs.FieldString("hash_key", "A optional hash key for partitioning messages.").IsInterpolated().Advanced(),
-			docs.FieldInt("max_in_flight", "The maximum number of parallel message batches to have in flight at any given time."),
-			policy.FieldSpec(),
-		).WithChildren(sess.FieldSpecs()...).WithChildren(retries.FieldSpecs()...).ChildDefaultAndTypesFromStruct(output.NewKinesisConfig()),
-		Categories: []string{
-			"Services",
-			"AWS",
-		},
-	})
-	if err != nil {
-		panic(err)
-	}
-}
-
 type kinesisWriter struct {
-	conf output.KinesisConfig
-
-	session *session.Session
+	conf    koConfig
 	kinesis kinesisiface.KinesisAPI
-
-	backoffCtor  func() backoff.BackOff
-	hashKey      *field.Expression
-	partitionKey *field.Expression
-	streamName   *string
-
-	log log.Modular
+	log     *service.Logger
 }
 
-func newKinesisWriter(conf output.KinesisConfig, mgr bundle.NewManagement) (*kinesisWriter, error) {
-	if conf.PartitionKey == "" {
-		return nil, errors.New("partition key must not be empty")
-	}
-
-	k := kinesisWriter{
-		conf:       conf,
-		log:        mgr.Logger(),
-		streamName: aws.String(conf.Stream),
-	}
-	var err error
-	if k.hashKey, err = mgr.BloblEnvironment().NewField(conf.HashKey); err != nil {
-		return nil, fmt.Errorf("failed to parse hash key expression: %v", err)
-	}
-	if k.partitionKey, err = mgr.BloblEnvironment().NewField(conf.PartitionKey); err != nil {
-		return nil, fmt.Errorf("failed to parse partition key expression: %v", err)
-	}
-	if k.backoffCtor, err = conf.Config.GetCtor(); err != nil {
-		return nil, err
-	}
-	return &k, nil
+func newKinesisWriter(conf koConfig, mgr *service.Resources) (*kinesisWriter, error) {
+	return &kinesisWriter{
+		conf: conf,
+		log:  mgr.Logger(),
+	}, nil
 }
 
 // toRecords converts an individual benthos message into a slice of Kinesis
@@ -118,16 +129,21 @@ func newKinesisWriter(conf output.KinesisConfig, mgr bundle.NewManagement) (*kin
 // and passing each new message through the partition and hash key interpolation
 // process, allowing the user to define the partition and hash key per message
 // part.
-func (a *kinesisWriter) toRecords(msg message.Batch) ([]*kinesis.PutRecordsRequestEntry, error) {
-	entries := make([]*kinesis.PutRecordsRequestEntry, msg.Len())
+func (a *kinesisWriter) toRecords(batch service.MessageBatch) ([]*kinesis.PutRecordsRequestEntry, error) {
+	entries := make([]*kinesis.PutRecordsRequestEntry, len(batch))
 
-	err := msg.Iter(func(i int, p *message.Part) error {
-		partKey, err := a.partitionKey.String(i, msg)
+	err := batch.WalkWithBatchedErrors(func(i int, m *service.Message) error {
+		partKey, err := batch.TryInterpolatedString(i, a.conf.PartitionKey)
 		if err != nil {
 			return fmt.Errorf("partition key interpolation error: %w", err)
 		}
+
+		mBytes, err := m.AsBytes()
+		if err != nil {
+			return err
+		}
 		entry := kinesis.PutRecordsRequestEntry{
-			Data:         p.AsBytes(),
+			Data:         mBytes,
 			PartitionKey: aws.String(partKey),
 		}
 
@@ -136,9 +152,11 @@ func (a *kinesisWriter) toRecords(msg message.Batch) ([]*kinesis.PutRecordsReque
 			return component.ErrMessageTooLarge
 		}
 
-		hashKey, err := a.hashKey.String(i, msg)
-		if err != nil {
-			return fmt.Errorf("hash key interpolation error: %w", err)
+		var hashKey string
+		if a.conf.HashKey != nil {
+			if hashKey, err = batch.TryInterpolatedString(i, a.conf.HashKey); err != nil {
+				return fmt.Errorf("hash key interpolation error: %w", err)
+			}
 		}
 		if hashKey != "" {
 			entry.ExplicitHashKey = aws.String(hashKey)
@@ -152,50 +170,44 @@ func (a *kinesisWriter) toRecords(msg message.Batch) ([]*kinesis.PutRecordsReque
 }
 
 func (a *kinesisWriter) Connect(ctx context.Context) error {
-	if a.session != nil {
+	if a.kinesis != nil {
 		return nil
 	}
 
-	sess, err := GetSessionFromConf(a.conf.SessionConfig.Config)
-	if err != nil {
-		return err
-	}
-
-	k := kinesis.New(sess)
+	k := kinesis.New(a.conf.session)
 	if _, err := k.DescribeStream(&kinesis.DescribeStreamInput{
-		StreamName: a.streamName,
+		StreamName: &a.conf.Stream,
 	}); err != nil {
 		return err
 	}
 
 	if err := k.WaitUntilStreamExists(&kinesis.DescribeStreamInput{
-		StreamName: a.streamName,
+		StreamName: &a.conf.Stream,
 	}); err != nil {
 		return err
 	}
 
-	a.session = sess
 	a.kinesis = k
 	a.log.Infof("Sending messages to Kinesis stream: %v\n", a.conf.Stream)
 
 	return nil
 }
 
-func (a *kinesisWriter) WriteBatch(ctx context.Context, msg message.Batch) error {
-	if a.session == nil {
+func (a *kinesisWriter) WriteBatch(ctx context.Context, batch service.MessageBatch) error {
+	if a.kinesis == nil {
 		return component.ErrNotConnected
 	}
 
-	backOff := a.backoffCtor()
+	backOff := a.conf.backoffCtor()
 
-	records, err := a.toRecords(msg)
+	records, err := a.toRecords(batch)
 	if err != nil {
 		return err
 	}
 
 	input := &kinesis.PutRecordsInput{
 		Records:    records,
-		StreamName: a.streamName,
+		StreamName: &a.conf.Stream,
 	}
 
 	// trim input record length to max kinesis batch size