+++ public/components/all/package.go
@@ -40,6 +40,7 @@ import (
 	_ "github.com/benthosdev/benthos/v4/public/components/pure/extended"
 	_ "github.com/benthosdev/benthos/v4/public/components/pusher"
 	_ "github.com/benthosdev/benthos/v4/public/components/redis"
+	_ "github.com/benthosdev/benthos/v4/public/components/sentry"
 	_ "github.com/benthosdev/benthos/v4/public/components/sftp"
 	_ "github.com/benthosdev/benthos/v4/public/components/snowflake"
 	_ "github.com/benthosdev/benthos/v4/public/components/splunk"